requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
numpy<2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain
langchain-openai
pymongo[srv]
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
datasets>=2.0.0